Move pubsub to demo cmd
