+++ publicroomsapi/storage/storage.go
@@ -20,10 +20,8 @@ import (
 	"net/url"
 
 	dht "github.com/libp2p/go-libp2p-kad-dht"
-	pubsub "github.com/libp2p/go-libp2p-pubsub"
 	"github.com/matrix-org/dendrite/publicroomsapi/storage/postgres"
 	"github.com/matrix-org/dendrite/publicroomsapi/storage/postgreswithdht"
-	"github.com/matrix-org/dendrite/publicroomsapi/storage/postgreswithpubsub"
 	"github.com/matrix-org/dendrite/publicroomsapi/storage/sqlite3"
 )
 
@@ -59,21 +57,5 @@ func NewPublicRoomsServerDatabaseWithDHT(dataSourceName string, dht *dht.IpfsDHT
 		return sqlite3.NewPublicRoomsServerDatabase(dataSourceName)
 	default:
 		return postgreswithdht.NewPublicRoomsServerDatabase(dataSourceName, dht)
-	}
-}
-
-// NewPublicRoomsServerDatabase opens a database connection.
-func NewPublicRoomsServerDatabaseWithPubSub(dataSourceName string, pubsub *pubsub.PubSub) (Database, error) {
-	uri, err := url.Parse(dataSourceName)
-	if err != nil {
-		return postgreswithpubsub.NewPublicRoomsServerDatabase(dataSourceName, pubsub)
-	}
-	switch uri.Scheme {
-	case schemePostgres:
-		return postgreswithpubsub.NewPublicRoomsServerDatabase(dataSourceName, pubsub)
-	case schemeFile:
-		return sqlite3.NewPublicRoomsServerDatabase(dataSourceName)
-	default:
-		return postgreswithpubsub.NewPublicRoomsServerDatabase(dataSourceName, pubsub)
 	}
 }
